bpf: rename UINT8_MAX to UINT16_max and fix cluster_id casts
The UINT8_MAX macro is defined as the 16bit hex value 0xFFFF. Since
cluster_id is now a uint16 to allow for extended Cluster Mesh, this
renames the macro to UINT16_MAX, rather than changing the value.

Lingering casts of cluster_id as a uint8 have also been corrected to
ensure its usage as uint16.

Fixes: e8752ec125fc ("bpf: Define UINT8_MAX")

+++ bpf/lib/eps.h
@@ -61,10 +61,10 @@ ipcache_lookup6(const void *map, const union v6addr *addr,
 	};
 
 	/* Check overflow */
-	if (cluster_id > UINT8_MAX)
+	if (cluster_id > UINT16_MAX)
 		return NULL;
 
-	key.cluster_id = (__u8)cluster_id;
+	key.cluster_id = (__u16)cluster_id;
 
 	ipv6_addr_clear_suffix(&key.ip6, prefix);
 	return map_lookup_elem(map, &key);
@@ -82,10 +82,10 @@ ipcache_lookup4(const void *map, __be32 addr, __u32 prefix, __u32 cluster_id)
 	};
 
 	/* Check overflow */
-	if (cluster_id > UINT8_MAX)
+	if (cluster_id > UINT16_MAX)
 		return NULL;
 
-	key.cluster_id = (__u8)cluster_id;
+	key.cluster_id = (__u16)cluster_id;
 
 	key.ip4 &= GET_PREFIX(prefix);
 	return map_lookup_elem(map, &key);
